Initial port to KF5.
Some things are also simplified, at least whatever I noticed.

Note that this plugin cannot be loaded as the .desktop file converted
to .json breaks the X-KDevelop-SupportedMimeTypes property. It then
contains both entries in a single entry, instead of two entries
with one mime each...

I'll wait for some suggestion on how to handle this.